+++ docs/changelog.rst
@@ -10,6 +10,8 @@ Development
 - Support MONGODB-AWS authentication mechanism (with `authmechanismproperties`) #2507
 - Turning off dereferencing for the results of distinct query. #2663
 - Add tests against Mongo 5.0 in pipeline
+- Bug fix support for PyMongo>=4 to fix "InvalidOperation: Cannot use MongoClient after close"
+  errors.
 
 Changes in 0.24.2
 =================